Partial format file example
