+++ java-tests-issue/src/main/java/miked/autoformat/AutoFormatIssue.java
@@ -8,12 +8,11 @@ class AutoFormatIssue {
     // @formatter:off
     // before auto format
     // https://stackoverflow.com/a/30026710/185123
-    static<T> CompletableFuture<List<T>> sequence(List<CompletableFuture<T>> com) {
+    static <T> CompletableFuture<List<T>> sequence(List<CompletableFuture<T>> com) {
         return CompletableFuture.allOf(com.toArray(new CompletableFuture<?>[0]))
                 .thenApply(v -> com.stream()
-                    .map(CompletableFuture::join)
-                    .collect(Collectors.toList())
-                );
+                        .map(CompletableFuture::join)
+                        .collect(Collectors.toList()));
     }
     // @formatter:on
