+++ java-tests-issue/src/main/java/miked/autoformat/SwithAutoFormatIssue.java
@@ -1,0 +1,27 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package miked.autoformat;
+
+public class SwithAutoFormatIssue {
+
+    // before format document ALT + SHIFT F
+    public String formatSwitchStatementBefore() {
+        switch ("hello") {
+            case "world":
+                return "ok";
+            default:
+                return "sure";
+        }
+    }
+
+    // after format document ALT + SHIFT F
+    public String formatSwitchStatementAfter() {
+        switch ("hello") {
+            case "world":
+                return "ok";
+            default:
+                return "sure";
+        }
+    }
+}
