+++ java-tests-issue/src/main/java/miked/javatestsissue/Library.java
@@ -8,24 +8,4 @@ public class Library {
         var id = "hello world";
         return true;
     }
-
-    // before format document ALT + SHIFT F
-    public String formatSwitchStatementBefore() {
-        switch ("hello") {
-            case "world":
-                return "ok";
-            default:
-                return "sure";
-        }
-    }
-
-    // after format document ALT + SHIFT F
-    public String formatSwitchStatementAfter() {
-        switch ("hello") {
-            case "world":
-                return "ok";
-            default:
-                return "sure";
-        }
-    }
 }
